curator/pyproject.toml
[tool.poetry]
name = "curator"
version = "0.1.0"
description = ""
authors = ["Christian Bonnell <christian.bonnell@gmail.com>"]
readme = "README.md"
include = ["resources/words/*.txt"]

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.3.0"
pydantic = "^2.7.2"
click = "^8.1.7"
tqdm = "^4.66.4"
google-api-python-client = "^2.131.0"
google-cloud-storage = "^2.16.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
rich = "^13.7.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
directory-import = "curator.directory_import:main"
search = "curator.search_and_download:main"
curator = "curator.cli:cli_group"
